pyproject.toml
[tool.poetry]
name = "digitallistening"
version = "0.1.0"
description = "Project template for data science projects"
authors = ["Juan Lara <julara@unal.edu.co>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.2"
matplotlib = "^3.5.0"
nltk = "^3.6.5"
wordcloud = "^1.8.1"
gensim = "^4.1.2"
emoji = "^1.6.1"
scikit-learn = "^1.0.1"
pydantic = "^1.8.2"
nbconvert = "^6.3.0"
pyarrow = "^6.0.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.2.4"
jupytext = "^1.13.2"
gdown = "^4.2.0"
pandas = "^1.3.4"
pandoc = "^2.0.1"
pyppeteer = "^0.2.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
